Moving rethinkdb new class into a separate file
